ECA-1439: certificate end validity date
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@8158 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/org/ejbca/core/model/ca/caadmin/CAInfo.java
@@ -76,7 +76,7 @@ public class CAInfo implements Serializable {
     protected String name;
     /** SecConst.CA_ACTIVE etc */
     protected int status;
-    protected int validity;
+    protected long validity;
     protected Date expiretime;
     protected Date updatetime;
     /** CATYPE_X509 or CATYPE_CVC */
@@ -115,7 +115,7 @@ public class CAInfo implements Serializable {
     public int getSignedBy() {return signedby;}
     public void setSignedBy(int signedby) {this.signedby = signedby;}
     
-    public int getValidity() { return validity;}
+    public long getValidity() { return validity;}
     public void setValidity(int validity) { this.validity = validity; }
     
     public Date getExpireTime() {return this.expiretime;}
